+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/ComposeTabbedBrowsingTest.kt
@@ -86,7 +86,7 @@ class ComposeTabbedBrowsingTest {
         mockWebServer.shutdown()
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903599
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/903599
     @Test
     fun closeAllTabsTest() {
         val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
@@ -118,7 +118,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903604
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/903604
     @Test
     fun closingTabsMethodsTest() {
         val genericURL = TestAssetHelper.getGenericAsset(mockWebServer, 1)
@@ -157,7 +157,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903591
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/903591
     @Test
     fun closingPrivateTabsMethodsTest() {
         val genericURL = TestAssetHelper.getGenericAsset(mockWebServer, 1)
@@ -199,7 +199,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903606
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/903606
     @SmokeTest
     @Test
     fun tabMediaControlButtonTest() {
@@ -240,6 +240,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/903602
     @Test
     fun verifyTabTrayNotShowingStateHalfExpanded() {
         homeScreen {
@@ -263,7 +264,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903600
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/903600
     @Test
     fun verifyEmptyTabTray() {
         homeScreen {
@@ -280,7 +281,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903585
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/903585
     @Test
     fun verifyEmptyPrivateTabsTrayTest() {
         homeScreen {
@@ -298,7 +299,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903601
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/903601
     @Test
     fun verifyTabsTrayWithOpenTabTest() {
         val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
@@ -322,7 +323,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903587
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/903587
     @SmokeTest
     @Test
     fun verifyPrivateTabsTrayWithOpenTabTest() {
@@ -347,7 +348,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/927314
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/927315
     @Test
     fun tabsCounterShortcutMenuTest() {
         val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
@@ -382,7 +383,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/927314
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/927314
     @Test
     fun privateTabsCounterShortcutMenuTest() {
         val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
@@ -418,7 +419,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/1046683
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/1046683
     @Test
     fun verifySyncedTabsWhenUserIsNotSignedInTest() {
         navigationToolbar {
@@ -432,7 +433,7 @@ class ComposeTabbedBrowsingTest {
         }
     }
 
-    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903598
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/903598
     @SmokeTest
     @Test
     fun shareTabsFromTabsTrayTest() {
