+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/DownloadTest.kt
@@ -210,28 +210,6 @@ class DownloadTest {
         deleteDownloadedFileOnStorage(downloadFile)
     }
 
-    // Save PDF file from the share overlay
-    @SmokeTest
-    @Test
-    fun saveAndOpenPdfTest() {
-        val genericURL =
-            TestAssetHelper.getGenericAsset(mockWebServer, 3)
-        downloadFile = "pdfForm.pdf"
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(genericURL.url) {
-            clickPageObject(itemWithText("PDF form file"))
-        }.openThreeDotMenu {
-        }.clickShareButton {
-        }.clickSaveAsPDF {
-            verifyDownloadPrompt(downloadFile)
-        }.clickDownload {
-        }.clickOpen("application/pdf") {
-            assertExternalAppOpens(GOOGLE_DOCS)
-        }
-        deleteDownloadedFileOnStorage(downloadFile)
-    }
-
     // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/1114970
     @Test
     fun deleteDownloadedFileTest() {
@@ -342,6 +320,7 @@ class DownloadTest {
         deleteDownloadedFileOnStorage(downloadFile)
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/457112
     @Ignore("Failing: https://bugzilla.mozilla.org/show_bug.cgi?id=1840994")
     @Test
     fun systemNotificationCantBeDismissedWhileInProgressTest() {
@@ -371,6 +350,7 @@ class DownloadTest {
         deleteDownloadedFileOnStorage(downloadFile)
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2299297
     @Test
     fun notificationCanBeDismissedIfDownloadIsInterruptedTest() {
         // Clear the "Firefox Fenix default browser notification"
@@ -404,6 +384,7 @@ class DownloadTest {
         deleteDownloadedFileOnStorage(downloadFile)
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/1632384
     @Test
     fun warningWhenClosingPrivateTabsWhileDownloadingTest() {
         downloadFile = "1GB.zip"
@@ -466,9 +447,11 @@ class DownloadTest {
         deleteDownloadedFileOnStorage(downloadFile)
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2048448
     // Save edited PDF file from the share overlay
+    @SmokeTest
     @Test
-    fun saveEditedPdfTest() {
+    fun saveAsPdfFunctionalityTest() {
         val genericURL =
             TestAssetHelper.getGenericAsset(mockWebServer, 3)
         downloadFile = "pdfForm.pdf"
